backendZeroCode/requirements.txt
flask
jsonify
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-flask
deprecated
